+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+- The solc v0.8.24 support with temporarily unsupported transient storage and blobs
+- The `MCOPY` instruction support
+- The `--evm-version` parameter to the CLI
 - An option to disable the system request memoization
 
 ## [1.3.23] - 2024-01-30